refactor (#266)
